modify DuiDesigner

+++ GameAndroidControlPCCoding/DuiLib/Core/UIMarkup.cpp
@@ -208,12 +208,12 @@ namespace DuiLib
         while (pstr < pstrEnd)
         {
             m_pOwner->_SkipWhitespace(pstr);
-            m_aAttributes[m_nAttributes].iName = pstr - m_pOwner->m_pstrXML;
+            m_aAttributes[m_nAttributes].iName = (ULONG)(pstr - m_pOwner->m_pstrXML);
             pstr += _tcslen(pstr) + 1;
             m_pOwner->_SkipWhitespace(pstr);
             if (*pstr++ != _T('\"')) return; // if( *pstr != _T('\"') ) { pstr = ::CharNext(pstr); return; }
 
-            m_aAttributes[m_nAttributes++].iValue = pstr - m_pOwner->m_pstrXML;
+            m_aAttributes[m_nAttributes++].iValue = (ULONG)(pstr - m_pOwner->m_pstrXML);
             if (m_nAttributes >= MAX_XML_ATTRIBUTES) return;
             pstr += _tcslen(pstr) + 1;
         }
@@ -493,8 +493,8 @@ namespace DuiLib
             _SkipWhitespace(pstrText);
             // Fill out element structure
             XMLELEMENT* pEl = _ReserveElement();
-            ULONG iPos = pEl - m_pElements;
-            pEl->iStart = pstrText - m_pstrXML;
+            ULONG iPos = (ULONG)(pEl - m_pElements);
+            pEl->iStart = (ULONG)(pstrText - m_pstrXML);
             pEl->iParent = iParent;
             pEl->iNext = pEl->iChild = 0;
             if (iPrevious != 0) m_pElements[iPrevious].iNext = iPos;
@@ -510,7 +510,7 @@ namespace DuiLib
             _SkipWhitespace(pstrText);
             if (pstrText[0] == _T('/') && pstrText[1] == _T('>'))
             {
-                pEl->iData = pstrText - m_pstrXML;
+                pEl->iData = (ULONG)(pstrText - m_pstrXML);
                 *pstrText = _T('\0');
                 pstrText += 2;
             }
@@ -518,7 +518,7 @@ namespace DuiLib
             {
                 if (*pstrText != _T('>')) return _Failed(_T("Expected start-tag closing"), pstrText);
                 // Parse node data
-                pEl->iData = ++pstrText - m_pstrXML;
+                pEl->iData = (ULONG)(++pstrText - m_pstrXML);
                 LPTSTR pstrDest = pstrText;
                 if (!_ParseData(pstrText, pstrDest, _T('<'))) return false;
                 // Determine type of next element
